Added platform and device info printing

+++ src/kernel_program.cl
@@ -3,7 +3,7 @@ __kernel void square(
 		     __global float* output,
 		     const unsigned int count)
 {
-  int i = get_global_id(0);
+  size_t i = get_global_id(0);
   if(i < count)
     output[i] = input[i] * input[i];
 }
